+++ backend/src/migration/migration_1_14_3.rs
@@ -1,12 +1,16 @@
+use crate::io::fs::basic_folder::{config_directory, home_directory};
+use std::fs::{copy, create_dir_all};
+use std::path::Path;
+
 pub fn migriere_1_14_3() {
-    let old_path = crate::io::fs::basic_folder::home_directory()
+    let old_path = home_directory()
         .join(".local")
         .join("share")
         .join("looksyk")
         .join("config.json");
-    let new_path = crate::io::fs::basic_folder::config_directory().join("config.json");
-    if std::path::Path::new(&old_path).exists() && !std::path::Path::new(&new_path).exists() {
-        std::fs::create_dir_all(crate::io::fs::basic_folder::config_directory()).unwrap();
-        std::fs::copy(&old_path, &new_path).unwrap();
+    let new_path = config_directory().join("config.json");
+    if Path::new(&old_path).exists() && !Path::new(&new_path).exists() {
+        create_dir_all(config_directory()).unwrap();
+        copy(&old_path, &new_path).unwrap();
     }
 }